Add toSqlLikeSuffixPattern function in expression.go
A new function, toSqlLikeSuffixPattern, has been added to handle creation of SQL "LIKE" pattern with a suffix wildcard. It can intake comparison values of type string and int, and is used in generating SQL WHERE clause conditions. Usage examples of this function can be seen in 'Sm' and 'Nsm' functions.