Update expression.go and add new tests in expression_test.go
Added new functions in expression.go to handle creation of SQL "LIKE" and "NOT LIKE" patterns. Updated several existing functions to improve readability and efficiency. Also, new unit tests have been added to expression_test.go to test the functionality of the new and updated methods. These changes improve the robustness and reliability of the code.