Add 'In' function and related tests in ExpressionSuite
Introduced a new function 'In' in the expression.go file which generates expressions based on a target and a list of values. Additional unit tests have been created in the expression_test.go file to ensure the correct working of the 'In' function. The function supports inputs of type string, []string, int, or []int.