Add IS/IS NOT TRUE/FALSE/NULL conditions and corresponding tests
Methods to create expressions for conditions like "IS NULL", "IS NOT NULL", "IS TRUE", "IS NOT TRUE", "IS FALSE" and "IS NOT FALSE" have been added in expression.go. The corresponding unit tests for these conditions have also been included in expression_test.go to ensure accuracy of the newly implemented methods.