Refactor code to accept any data type as function parameter
The code has been updated to allow the parameters of all functions in 'expresssion.go' to be of any data type, previously only accepted strings. This greatly extends the flexibility and utility of these functions as they can now handle wider variety of data types. Appropriate conversions are handled within the functions themselves to maintain the integrity of the operations.